sched: scale the busy and this queue's per-task load before compare
Date	Mon, 17 Jun 2013 21:00:23 +0800

Since for max_load and this_load, they are the value that already be
scaled. It is not reasonble to get a minimum value between the scaled
and non-scaled value, like below example.
	min(sds->busiest_load_per_task, sds->max_load);

Also add comment over in what condition, there would be cpu power gain
in move the load.

Change-Id: I6c11f2782862f9b257ca0428ea4fed728694bc67
Signed-off-by: Lei Wen <leiwen@marvell.com>